requirements.txt
tensorflow>=2.12
flask
numpy
pillow
scikit-learn
matplotlib
gunicorn
